feat: :sparkles: categorySlice 작성
