--- src/store/setSelectedCategory.tsx
@@ -1,22 +1,0 @@
-import { createSlice, PayloadAction } from '@reduxjs/toolkit'
-
-interface CategoryState {
-  selectedCategory: string | null
-}
-
-const initialState: CategoryState = {
-  selectedCategory: null,
-}
-
-const categorySlice = createSlice({
-  name: 'category',
-  initialState,
-  reducers: {
-    setSelectedCategory: (state, action: PayloadAction<string | null>) => {
-      state.selectedCategory = action.payload
-    },
-  },
-})
-
-export const { setSelectedCategory } = categorySlice.actions
-export default categorySlice.reducer
